Initial commit of patient context picker
In the form of an Authenticator, this picker can be used for custom auth
flows in order to ask the user to select a single patient context.
The authenticator uses a ResteasyClient to make requests against a
configured FHIR Server to obtain information like name and date-of-birth
for each patient resource id.