Refactor
1. Use a batch read request instead of a search since we already know
the id of each resource that we want to get

2. Set the patient/Patient.read scope

3. Set a patient_id claim in the internal access token for the FHIR
resource server to use to enforce authorization (so we only get
information on the context picker that the end user is allowed to see).
    Note:  we should probably use `user/Patient.read` here instead since
we want to support the read beyond a single patient scope.

4. Avoid getting the patient info if there is only a single resourceId
to use for context

5. Avoid presenting the patient context picker if there is only a single
patient returned (in case we have multiple resourceId values, but only
one is on the server)

6. Use the IBM FHIR Server's JAX-RS provider for application/fhir+json
and the com.ibm.fhir:fhir-model for improved ergonomics while working
with the FHIR request